dom/l10n/tests/mochitest/dom_localization/test_overlay.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test DOMLocalization's DOMOverlay functionality</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript">
  "use strict";
  const l10nReg = new L10nRegistry();
  const fs = [
    { path: "/localization/en-US/mock.ftl", source: `
title = <strong>Hello</strong> World
title2 = This is <a data-l10n-name="link">a link</a>!
` },
  ];
  const source = L10nFileSource.createMock("test", "app", ["en-US"], "/localization/{locale}", fs);
  l10nReg.registerSources([source]);

  window.onload = async function() {
    SimpleTest.waitForExplicitFinish();

    const domLoc = new DOMLocalization(
      ["/mock.ftl"],
      false,
      l10nReg,
      ["en-US"],
    );

    const p1 = document.querySelectorAll("p")[0];
    const p2 = document.querySelectorAll("p")[1];
    const a = p2.querySelector("a");
    // We want to test that the event listener persists after
    // translateFragment().
    a.addEventListener("click", function(e) {
      SimpleTest.finish();
      // We cannot connect to non-local connections on automation, so prevent
      // the navigation.
      e.preventDefault();
    });

    await domLoc.translateFragment(document.body);


    is(p1.querySelector("strong").textContent, "Hello");

    is(p2.querySelector("a").getAttribute("href"), "http://www.mozilla.org");
    is(p2.querySelector("a").textContent, "a link");

    a.click();
  };
  </script>
</head>
<body>
  <p data-l10n-id="title" />
  <p data-l10n-id="title2">
    <a href="http://www.mozilla.org" data-l10n-name="link"></a>
  </p>
</body>
</html>
